added new Screens for travel planning
