+++ Travel/OverviewTravelPlanningTVC.swift
@@ -1,46 +1,34 @@
 //
-//  TravelPlanningTableViewController.swift
+//  OverviewTravelPlanningTVC.swift
 //  Travel
 //
-//  Created by Jessica Ernst on 09.09.22.
+//  Created by Jessica Ernst on 16.09.22.
 //
 
 import UIKit
 
-class TravelPlanningTableViewController: UITableViewController {
+class OverviewTravelPlanningTVC: UITableViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        // Add a background view to the table view
-        let backgroundImage = UIImage(named: "dariusz-sankowski-3OiYMgDKJ6k-unsplash")
-        let imageView = UIImageView(image: backgroundImage)
-        imageView.contentMode = .scaleAspectFill
-        let blurEffect = UIBlurEffect(style: UIBlurEffect.Style.regular)
-        let blurView = UIVisualEffectView(effect: blurEffect)
-        blurView.frame = imageView.bounds
-        imageView.addSubview(blurView)
-        self.tableView.backgroundView = imageView
-        
-        let appearance = UINavigationBarAppearance()
-        appearance.configureWithOpaqueBackground()
-        appearance.backgroundColor = UIColor.clear
-                
-        let navigationBar = self.navigationController!.navigationBar
-        navigationBar.standardAppearance = appearance
+        // Uncomment the following line to preserve selection between presentations
+        // self.clearsSelectionOnViewWillAppear = false
+
+        // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
+        // self.navigationItem.rightBarButtonItem = self.editButtonItem
     }
-    
 
     // MARK: - Table view data source
 
     override func numberOfSections(in tableView: UITableView) -> Int {
-        
-        return 1
+        // #warning Incomplete implementation, return the number of sections
+        return 0
     }
 
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        
-        return 6
+        // #warning Incomplete implementation, return the number of rows
+        return 0
     }
 
     /*
